refactor: extract pipeline domain to pipeline-store
Phase 6 of god store decomposition. Moves pipeline configuration
(per-project pipeline config, step CRUD, reordering) to a dedicated
usePipelineStore. Subscription syncs state back to app-store.

app-store: 2,254 → 2,172 lines (-82)
